chore(db): rename entity files
